Refactor: Extract class
